Move QuizGenerator into business module
