+++ app/src/main/java/it/ilker/repeatcard/ui/quiz/QuizViewModel.kt
@@ -5,13 +5,14 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import it.ilker.repeatcard.db.FlashcardDatabase
 import it.ilker.repeatcard.db.flashcard.FlashcardRepository
-import it.ilker.repeatcard.models.question.Question
-import it.ilker.repeatcard.models.quizresult.QuizResult
+import me.ilker.business.question.Question
+import me.ilker.business.quiz.QuizResult
 import it.ilker.repeatcard.ui.util.exhaustive
 import java.util.UUID
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
+import me.ilker.business.quiz.QuestionGenerator
 
 const val MIN_CARD_NUMBER_FOR_QUIZ = 4
