+++ app/src/main/java/it/ilker/repeatcard/ui/results/ResultsScreen.kt
@@ -11,8 +11,8 @@ import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import it.ilker.repeatcard.R
-import it.ilker.repeatcard.models.question.Question
-import it.ilker.repeatcard.models.quizresult.QuizResult
+import me.ilker.business.question.Question
+import me.ilker.business.quiz.QuizResult
 import it.ilker.repeatcard.ui.question.QuestionDetailScreen
 import it.ilker.repeatcard.ui.util.exhaustive
 import kotlinx.android.synthetic.main.screen_results.*
