+++ app/src/test/java/it/ilker/repeatcard/ui/quiz/QuestionGeneratorTest.kt
@@ -2,7 +2,7 @@ package it.ilker.repeatcard.ui.quiz
 
 import org.junit.Test
 
-import it.ilker.repeatcard.models.question.Question
+import me.ilker.business.question.Question
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
